+++ CHANGES.md
@@ -3,13 +3,14 @@
 ## v1.2.0
 #### 2015-XXX-XX
 - Bugfix (show_index,...)
+- Added extra configuration option for column width
 - Added optional user/password from command line (using --user <user> and --password <password>)
 - Added size info to mytables command
 - Added append_tables option to append data to existing tables
 - Minor changes in uploading tables
 - **python 3 compatible**
 - **change config.ini folder, from .easyacess/ to .easyaccess (typo)**
-- move multiprocess import depending on loading bar
+- move multiprocess import call depending on loading bar is set
 
 ## v1.1.0
 #### 2015-APR-28